requirements.txt
Flask==2.0.1
FPDF2==2.4.3